Add smoke tests (#17)
I'm about to upgrade the runtime dependencies for this project. First, I wanted to create a sample app that I could run smoke-tests with.

By having these smoke tests, I can be more confident in the upgrade.

## Approach
I'm using a docker container for the redis instance, seeding it with some default data and then testing end-to-end with `curl`.